Add simpler self serve ConfigWriter
